Move CortxClusterManager method to cluster controller
