+++ ha/const.py
@@ -33,8 +33,6 @@ RULE_GLOBAL_INDEX="rules_engine"
 HA_CONFIG_FILE="{}/ha.conf".format(CONFIG_DIR)
 HA_GLOBAL_INDEX="ha_conf"
 SOURCE_CONFIG_FILE="{}/ha.conf".format(SOURCE_CONFIG_PATH)
-CLUSTER_MANAGER_CONTROLLER_INDEX="controller_interface"
-CLUSTER_MANAGER_CONTROLLER_SCHEMA="{}/controller_interface.json".format(CONFIG_DIR)
 
 CURRENT_NODE_STATUS="self_node_status"
 OTHER_NODE_STATUS="other_node_status"
@@ -92,3 +90,8 @@ PCS_CLUSTER_START="pcs cluster start --all"
 PCS_CLUSTER_STATUS="pcs cluster status"
 PCS_CLUSTER_UNSTANDBY="pcs cluster unstandby --all"
 PCS_STATUS_NODES="pcs status nodes"
+
+# Cluster manager
+CLUSTER_MANAGER_CONTROLLER_INDEX="controller_interface"
+CLUSTER_MANAGER_CONTROLLER_SCHEMA="{}/controller_interface.json".format(CONFIG_DIR)
+CLUSTER_MANAGER_ELEMENT=["cluster", "node", "service"]
